Allow 'common-fate-allow-assume-role' tag to be used for trust policies
Our S3 Audit Log Destination requires that Common Fate assumes a role to write to the bucket.

Because we use ABAC to control access to these roles, a user would have to add a 'common-fate-integration-read-role' to the S3 write role.
This is confusing, because we're writing rather than reading.

This commit adds a new tag 'common-fate-allow-assume-role' which can be used instead, and is less confusing.